fixed save file issue
use to save the output file in current directory, now looks for output file in the same directory as the executable. Not in love with the way I did this, but it works.
I think I'm going to do a rewrite of this program so I can port it over to android